-Rollback on scanner
