+++ server/src/AnalysisManager.ts
@@ -61,10 +61,10 @@ export class AnalysisManager{
         // checks with table
 
         const wellformednessDiagnostics = await checks_with_table(symbolTable, document, tree.rootNode,this.documentSymbolTables);
-        //console.log("Symbol table created for:", document.uri);
+        console.log("Symbol table created for:", document.uri);
         console.log("Number of symbols found:", symbolTable.getSymbols().length);
         console.log("Wellformedness diagnostics:", wellformednessDiagnostics.length);
-        //console.log("Symbol table", symbolTable);
+        console.log("Symbol table", symbolTable);
         const allDiagnostics = [
             ...syntaxDiagnostics,
             ...wellformednessDiagnostics,
